Initial (highly derivative) MS Lync processor/recipe checkin, updates
Since this is practically a 100% subset of the MSOffice2011_updates
minus the URL and the name/info.plist, it’s almost an exact facsimile
of the core autopkg/recipes for Office, with different blocking
applications for the munki recipe.
In the process of converting all the other recipes to a more modern
format.