Updates 8x8 Work recipes
The prior download url seems to have moved or now be rendered via javascript.

I dug around and found a location that contains both intel and arm download URL's.

With this, and two new input variables, the `*-Intel` recipes can go.

I added a `StopProcessingIf` to the Munki recipe to make sure that folks set `MUNKI_ARCHITECTURE` to either `arm64` or `x86_64`.

Sadly the processor raises `NSInvalidArgumentException` is `MUNKI_ARCHITECTURE` is not set (https://macadmins.slack.com/archives/C056155B4/p1716994896073719).

But the defaults are set in a way that arm is the default.

Instead of deleting the `*-Intel` recipes, I can amend to add `DeprecationWarning` to alert folks to the changes if wanted.